(#503) Make hide-goal and show-goal accessible in console
